Refactor platform-specifics to be clearer about: - What is platform-specific - Memory allocations incurred when creating platform-specific elements

+++ source/math/Rect.cpp
@@ -7,7 +7,7 @@ Rect::Rect()
     , height(0)
 {}
 
-Rect::Rect(int x, int y, int width, int height)
+Rect::Rect(float x, float y, float width, float height)
     : x(x)
     , y(y)
     , width(width)
